migrated routes to controllers, adjusted for new file structure. more modules
